% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.r
\name{aggregate_country_data}
\alias{aggregate_country_data}
\title{Aggregate country data}
\usage{
aggregate_country_data(data, agg_columns, group_columns)
}
\arguments{
\item{data}{Dataframe with data for all years of a country in a given HS classification. Dataframe should be raw output from Comtrade}

\item{agg_columns}{list of column names to be aggregated/summed up}

\item{group_columns}{list of columns to group the aggregation by}
}
\value{
A \code{dataframe} containing the same columns as the input dataframe, less those not either aggregation columns or group by columns
}
\description{
Combine all the years of data for a country in a particular HS classification. Purpose is to have a more robust distribution for 1->n conversions for \code{covert_hs} function, \code{map_df} parameter.
}
