% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.r
\name{aggregate_digit_level}
\alias{aggregate_digit_level}
\title{Aggregate comtrade data to a different digit level}
\usage{
aggregate_digit_level(
  df,
  agg_columns,
  group_columns,
  commodity_column,
  digit_level
)
}
\arguments{
\item{df}{The dataframe to aggregate}

\item{agg_columns}{list of column names to be aggregated/summed up}

\item{group_columns}{list of columns to group the aggregation by}

\item{commodity_column}{name of column with commodity codes}

\item{digit_level}{number of digits to aggregate to, e.g., '6' for 6-digit, etc.}
}
\value{
A \code{dataframe}
}
\description{
Aggregate comtrade data to a different digit level.
}
