% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_hs.r
\name{convert_hs}
\alias{convert_hs}
\title{Convert Comtrade data from one HS year to another}
\usage{
convert_hs(correspondence_table, hs_from, hs_to, df, map_df = NA)
}
\arguments{
\item{correspondence_table}{Dataframe with HS mappings, obtained from the \code{get_correspondence_tables} function.}

\item{hs_from}{Integer of which HS year the original data is in (e.g., '2017').}

\item{hs_to}{Integer of which HS year the data should be converted to (e.g., '2017').}

\item{df}{The original dataframe to convert.}

\item{df_map}{A dataframe of the country in the desired HS year to use for 1->n mappings. If omitted, will use equal distribution for 1-n mappings.}
}
\value{
A \code{dataframe} containing the following columns:

\item{first}{the 'to' product code}
\item{second}{the 'from' product code}
\item{first_name}{the hs version of the 'to' product code}
\item{second_name}{the hs version of the 'from' product code}
\item{relationship}{the correspondence type of this product pair, e.g. 1:n, n:n, etc.}
}
\description{
Convert Comtrade data from one HS year to another
}
