% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_correspondence_tables.r
\name{convert_hs_name}
\alias{convert_hs_name}
\title{Convert HS classification names to years}
\usage{
convert_hs_name(hs_name)
}
\value{
A number or string

\item{hs_name}{the HS name or year}
}
\description{
Convert HS classification names to years (e.g., "H0" > 1992) or vice versa
}
