% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_correspondence_tables.R
\name{get_correspondence_tables}
\alias{get_correspondence_tables}
\title{Get correspondence tables from UN stat}
\usage{
get_correspondence_tables()
}
\value{
A \code{dataframe} containing the following columns:

\item{first}{the 'to' product code}
\item{second}{the 'from' product code}
\item{first_name}{the hs version of the 'to' product code}
\item{second_name}{the hs version of the 'from' product code}
\item{relationship}{the correspondence type of this product pair, e.g. 1:n, n:n, etc.}
}
\description{
Gets all correspondence tables from the UN and generates a CSV with all the mappings. To avoid pinging the UN's servers too often, the output can be saved to a CSV and reread from there.
}
