% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_country_data.R
\name{aggregate_country_data}
\alias{aggregate_country_data}
\title{Aggregate country data}
\usage{
aggregate_country_data(data)
}
\arguments{
\item{data}{Dataframe with data for all years of a country in a given HS classification. Dataframe should be raw output from Comtrade with columns:

\item{Year}{the year of the trade data}
\item{FlowCode}{the HS version of the data}
\item{ReporterCode}{code of the trade flow (export, import, etc.)}
\item{ReporterLabel}{the name of the reporter country}
\item{PartnerCode}{the code of the partner country}
\item{CommodityCode}{the commodity code}
\item{PartnerLabel}{the name of the partner country}
\item{Value}{the value of recorded trade}}
}
\value{
A \code{dataframe} containing the following same columns as the input dataframe, with \code{Year} set to 999 to denote it is an aggregation, but to make sure output df has the same columns as the input df.
}
\description{
Combine all the years of data for a country in a particular HS classification. Purpose is to have a more robust distribution for 1->n conversions for \code{covert_hs} function, \code{map_df} parameter.
}
