% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_hs.R
\name{convert_hs}
\alias{convert_hs}
\title{Convert Comtrade data from one HS year to another}
\usage{
convert_hs(correspondence_table, hs_from, hs_to, df, map_df = NA)
}
\arguments{
\item{correspondence_table}{Dataframe with HS mappings, obtained from the \code{get_correspondence_tables} function.}

\item{hs_from}{Integer of which HS year the original data is in (e.g., '2017').}

\item{hs_to}{Integer of which HS year the data should be converted to (e.g., '2017').}

\item{df}{The original dataframe to convert with columns:

\itemize{
\item{Year}{the year of the trade data}
\item{FlowCode}{the HS version of the data}
\item{ReporterCode}{code of the trade flow (export, import, etc.)}
\item{ReporterLabel}{the name of the reporter country}
\item{PartnerCode}{the code of the partner country}
\item{PartnerLabel}{the name of the partner country}
\item{CommodityCode}{the commodity code}
\item{Value}{the value of recorded trade}}}

\item{map_df}{A dataframe of the country in the desired HS year to use for 1->n mappings. If omitted, will use equal distribution for 1-n mappings. Same columns as \code{df} parameter dataframe.}
}
\value{
A \code{dataframe} containing the following columns:

\itemize{
\item{hs_version}{the HS version of the data} 
\item{year}{the year of the trade data}
\item{flow_code}{code of the trade flow (export, import, etc.)}
\item{flow}{the name of the trade flow}
\item{reporter_code}{the code of the reporter country}
\item{reporter}{the name of the reporter country}
\item{partner_code}{the code of the partner country}
\item{partner}{the name of the partner country}
\item{commodity_code}{the commodity code}
\item{value}{the value of recorded trade}}
}
\description{
Convert Comtrade data from one HS year to another. Will additionally add a "World" aggregate. Note that the function may take some time to run even on small datasets due to the complex logic necessary for n:n transformations.
}
