% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_hs.r
\name{convert_hs}
\alias{convert_hs}
\title{Convert Comtrade data from one HS year to another}
\usage{
convert_hs(
  correspondence_tables,
  hs_from,
  hs_to,
  df,
  agg_columns,
  group_columns,
  commodity_column,
  aggregate_order = NA,
  map_df = NA,
  quiet = TRUE
)
}
\arguments{
\item{correspondence_tables}{Dataframe with HS mappings, obtained from the \code{get_correspondence_tables} function.}

\item{hs_from}{Integer of which HS year the original data is in (e.g., '2017').}

\item{hs_to}{Integer of which HS year the data should be converted to (e.g., '2017').}

\item{df}{The original dataframe to convert with columns}

\item{agg_columns}{list of column names to be aggregated/summed up}

\item{group_columns}{list of columns to group the aggregation by}

\item{commodity_column}{name of column with commodity codes}

\item{aggregate_order}{order of aggregation/values columns to use in mapping if one is missing in the mapping dataframe}

\item{map_df}{A dataframe of the country in the desired HS year to use for 1->n mappings. If omitted, will use equal distribution for 1-n mappings. Same columns as \code{df} parameter dataframe.}

\item{quiet}{A boolean of whether or not to print out progress}
}
\value{
A \code{dataframe}
}
\description{
Convert Comtrade data from one HS year to another.Not that any commodity codes that are not made up exclusively of numbers (e.g., a value like "TOTAL"), will be dropped from the conversion
}
