% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_hs.r
\name{convert_hs}
\alias{convert_hs}
\title{Convert Comtrade data from one HS year to another}
\usage{
convert_hs(correspondence_table, hs_from, hs_to, df, map_df = NA)
}
\arguments{
\item{correspondence_table}{Dataframe with HS mappings, obtained from the \code{get_correspondence_tables} function.}

\item{hs_from}{Integer of which HS year the original data is in (e.g., '2017').}

\item{hs_to}{Integer of which HS year the data should be converted to (e.g., '2017').}

\item{df}{The original dataframe to convert.}

\item{map_df}{A dataframe of the country in the desired HS year to use for 1->n mappings. If omitted, will use equal distribution for 1-n mappings.}
}
\value{
A \code{dataframe} containing the following columns:

\itemize{
\item{hs_version}{the HS version of the data} 
\item{year}{the year of the trade data}
\item{flow_code}{code of the trade flow (export, import, etc.)}
\item{flow}{the name of the trade flow}
\item{reporter_code}{the code of the reporter country}
\item{reporter}{the name of the reporter country}
\item{partner_code}{the code of the partner country}
\item{partner}{the name of the partner country}
\item{commodity_code}{the commodity code}
\item{value}{the value of recorded trade}}
}
\description{
Convert Comtrade data from one HS year to another
}
